pyproject.toml
[tool.poetry]
name = "bone-fracture-detection"
version = "0.1.0"
description = ""
authors = ["Ewen Le Bihan <hey@ewen.works>"]

[tool.poetry.dependencies]
python = "^3.9,<3.10"
scikit-image = "^0.18.1"
opencv-python = "^4.5.1"
matplotlib = "3.3.4"
docopt = "^0.6.2"
svgpathtools = "^1.4.1"
PyYAML = "^5.4.1"
keras = "^2.6.0"
beautifulsoup4 = "^4.9.3"
python-slugify = "^5.0.2"
helium = "^3.0.7"
tensorflow = "2.5"
nptyping = "^1.4.4"
keras-rl = "^0.4.2"
Pillow = "^9.0.1"
deco = "^0.6.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.4b2", allow-prereleases = true}
rich = "^10.1.0"
rope = "^0.19.0"
ipython = "^7.24.1"
pylint = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
